app/__init__.py
from flask import Flask
from extensions import db, login_manager
from flask_migrate import Migrate
from config import Config

app = Flask(__name__)
app.config.from_object(Config)

db.init_app(app)
login_manager.init_app(app)
login_manager.login_view = 'auth.login'

migrate = Migrate(app, db)

# Импорт и регистрация Blueprint'ов
from auth.routes import auth_bp
from inventory.routes import inventory_bp

app.register_blueprint(auth_bp)
app.register_blueprint(inventory_bp)
